requirements.txt
google-generativeai
python-dotenv
pytrends
backoff
